Prompt on destructive user actions

+++ apps/settings/src/components/UserList/UserRow.vue
@@ -1,7 +1,9 @@
 <!--
   - @copyright Copyright (c) 2018 John Molakvoæ <skjnldsv@protonmail.com>
+  - @copyright Copyright (c) 2019 Gary Kim <gary@garykim.dev>
   -
   - @author John Molakvoæ <skjnldsv@protonmail.com>
+  - @author Gary Kim <gary@garykim.dev>
   -
   - @license GNU AGPL version 3 or any later version
   -
@@ -428,25 +430,65 @@ export default {
 		},
 
 		wipeUserDevices() {
-			this.loading.wipe = true
-			this.loading.all = true
 			let userid = this.user.id
-			return this.$store.dispatch('wipeUserDevices', userid)
-				.then(() => {
-					this.loading.wipe = false
-					this.loading.all = false
-				})
+			OC.dialogs.prompt(
+				t('settings', 'In case of lost device or exiting the organization, this can remotely wipe the Nextcloud data from all devices associated with the account. Only works if the devices are connected to the internet. Type "{userid}" to confirm', { userid: userid }),
+				t('settings', 'Remote wipe of devices'),
+				(result, input) => {
+					if (!result) {
+						return
+					}
+					if (input === userid) {
+						this.loading.wipe = true
+						this.loading.all = true
+						this.$store.dispatch('wipeUserDevices', userid)
+							.then(() => {
+								this.loading.wipe = false
+								this.loading.all = false
+							})
+					} else {
+						OCP.Toast.error(t('settings', 'Username incorrect'))
+					}
+				},
+				true,
+				t('settings', 'Username'),
+				false,
+				{
+					confirm: t('settings', 'Wipe {userid}\'s Devices', { userid: userid }),
+					cancel: t('settings', 'Cancel')
+				}
+			)
 		},
 
 		deleteUser() {
-			this.loading.delete = true
-			this.loading.all = true
 			let userid = this.user.id
-			return this.$store.dispatch('deleteUser', userid)
-				.then(() => {
-					this.loading.delete = false
-					this.loading.all = false
-				})
+			OC.dialogs.prompt(
+				t('settings', 'Fully delete the user account including all their personal files, app data, etc. Type "{userid}" to confirm.', { userid: userid }),
+				t('settings', 'Account Deletion'),
+				(result, input) => {
+					if (!result) {
+						return
+					}
+					if (input === userid) {
+						this.loading.delete = true
+						this.loading.all = true
+						return this.$store.dispatch('deleteUser', userid)
+							.then(() => {
+								this.loading.delete = false
+								this.loading.all = false
+							})
+					} else {
+						OCP.Toast.error(t('settings', 'Username incorrect'))
+					}
+				},
+				true,
+				t('settings', 'Username'),
+				false,
+				{
+					confirm: t('settings', 'Delete {userid}\'s account', { userid: userid }),
+					cancel: t('settings', 'Cancel')
+				}
+			)
 		},
 
 		enableDisableUser() {
